% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lubridate.R
\name{firstOfMonth}
\alias{firstOfMonth}
\title{Function to get the first of the month.}
\usage{
firstOfMonth(date)
}
\arguments{
\item{date}{vector of dates}
}
\value{
vector of dates containing the corresponding mondays
}
\description{
Function to get the first of the month.
}
\note{
Function could have been written with pure strptime commands in case
one does not want the lubridate dependencies.
}

