% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{get.incidence2}
\alias{get.incidence2}
\title{Overwrites the get.incidence function for obkData objects to support doBy}
\usage{
get.incidence2(x, ...)
}
\arguments{
\item{x}{The object (typically obkdata) to compute the incidence for.}

\item{...}{Whatever}
}
\description{
Creates different time series based on the list of factor variables.
This function should eventually migrate back into the OutbreakTools package.
}
\author{
Michael Höhle
}

