% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocument.R
\docType{data}
\name{hagelloch.obk}
\alias{hagelloch.obk}
\title{Converted Hagelloch 1861 measles outbreak into obkData format.}
\format{\preformatted{Formal class 'obkData' [package "OutbreakTools"] with 6 slots
  ..@ individuals:'data.frame':	187 obs. of  25 variables:
  .. ..$ NAME : Factor w/ 27 levels "Aberle","Buerker",..: 15 15 15 23 14 14 21 21 21 21 ...
  .. ..$ FN   : int [1:187] 41 41 41 61 42 42 26 44 44 29 ...
  .. ..$ HN   : int [1:187] 61 61 61 62 63 63 23 69 69 31 ...
  .. ..$ AGE  : num [1:187] 7 6 4 13 8 12 6 10 13 7 ...
  .. ..$ SEX  : Factor w/ 2 levels "male","female": 2 2 2 1 2 1 1 1 1 2 ...
  .. ..$ PRO  : Date[1:187], format: "1861-11-21" "1861-11-23" ...
  .. ..$ ERU  : Date[1:187], format: "1861-11-25" "1861-11-27" ...
  .. ..$ CL   : Factor w/ 3 levels "preschool","1st class",..: 2 2 1 3 2 3 1 2 3 2 ...
  .. ..$ DEAD : Date[1:187], format: NA NA ...
  .. ..$ IFTO : int [1:187] 45 45 172 180 45 180 42 45 182 45 ...
  .. ..$ SI   : int [1:187] 10 12 9 10 11 9 9 10 11 10 ...
  .. ..$ C    : Factor w/ 6 levels "no complicatons",..: 1 1 1 1 1 2 1 2 1 2 ...
  .. ..$ PR   : int [1:187] 4 4 4 1 5 3 4 5 4 4 ...
  .. ..$ CA   : int [1:187] 4 4 4 1 3 3 3 3 3 5 ...
  .. ..$ NI   : int [1:187] 3 3 3 1 2 2 1 2 2 1 ...
  .. ..$ GE   : int [1:187] 1 1 2 1 1 1 1 1 1 1 ...
  .. ..$ TD   : int [1:187] NA 3 1 3 NA 2 2 6 1 11 ...
  .. ..$ TM   : num [1:187] NA 40.3 40.5 40.7 NA 40.7 39.3 39.9 40.7 38.4 ...
  .. ..$ x.loc: num [1:187] 142 142 142 165 145 ...
  .. ..$ y.loc: num [1:187] 100 100 100 102 120 ...
  .. ..$ tPRO : num [1:187] 22.7 24.2 29.6 28.1 23.1 ...
  .. ..$ tERU : num [1:187] 26.2 28.8 33.7 29 28.4 ...
  .. ..$ tDEAD: num [1:187] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ tR   : num [1:187] 29.2 31.8 36.7 32 31.4 ...
  .. ..$ tI   : num [1:187] 21.7 23.2 28.6 27.1 22.1 ...
  ..@ records    :List of 3
  .. ..$ timePRO :'data.frame':	187 obs. of  2 variables:
  .. .. ..$ individualID: chr [1:187] "1" "2" "3" "4" ...
  .. .. ..$ date        : Date[1:187], format: "1861-11-21" ...
  .. ..$ timeERU :'data.frame':	187 obs. of  2 variables:
  .. .. ..$ individualID: chr [1:187] "1" "2" "3" "4" ...
  .. .. ..$ date        : Date[1:187], format: "1861-11-25" ...
  .. ..$ timeDEAD:'data.frame':	187 obs. of  2 variables:
  .. .. ..$ individualID: chr [1:187] "1" "2" "3" "4" ...
  .. .. ..$ date        : Date[1:187], format: NA ...
  ..@ dna        : NULL
  ..@ contacts   : NULL
  ..@ context    : NULL
  ..@ trees      : NULL
}}
\usage{
hagelloch.obk
}
\description{
Small converter function to take the hagelloch.df data.frame
available in the surveillance package and make an obkData object
out of it.
}
\details{
This function is not really one you would put in a package.
Instead, the function would be called for it's output. See the surveillance package
for a description of the data.
}
\author{
Michael Höhle <http://www.math.su.se/~hoehle>
}
\keyword{data}

