% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{inc2xts}
\alias{inc2xts}
\title{Helper function to format a get.incidence list of data.frames
to a multivariate xts object}
\usage{
inc2xts(incList)
}
\arguments{
\item{incList}{List of lists containing the data.frames from get.incidence2}
}
\value{
An xts object corresponding to the flattened incList
}
\description{
Helper function to format a get.incidence list of data.frames
to a multivariate xts object
}

