% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getIncidence.R
\name{linelist2xts}
\alias{linelist2xts}
\title{Simple aggregator based on the data.table function}
\usage{
linelist2xts(formula, dateProjFun = identity, dRange = NULL, data)
}
\arguments{
\item{formula}{Formula of type LHSVar ~ RHSVar, where LHSVar is
a column of class Dates and RHSVar is a column containing a factor variable}

\item{dateProjFun}{Date projection function, i.e. how are dates projected down (i.e. monday, start of month, etc.). Result must be a Date again!}

\item{dRange}{Vector of length two giving the start and stop dates, the actual range will be range(dateProjFun(dRange)).}

\item{data}{data.frame containing the LHSVar RHSVar variables.}
}
\value{
xts time series
}
\description{
Simple aggregator based on the data.table function
}

