% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lubridate.R
\name{monday}
\alias{monday}
\title{Function to get the monday of the corresponding week a date belongs to.}
\usage{
monday(date)
}
\arguments{
\item{date}{vector of dates}
}
\value{
vector of dates
}
\description{
Function to get the monday of the corresponding week a date belongs to.
}
\note{
Function could have been written with pure strptime commands in case
one does not want the lubridate dependencies. This is the function Sebastian
always wanted, sorry it's not in a standalone package.
}

