% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lubridate.R
\name{wdaymon}
\alias{wdaymon}
\title{Version of the wday function from lubridate, where the monday is the 1.}
\usage{
wdaymon(date)
}
\arguments{
\item{date}{- vector of dates}
}
\value{
vector of numeric with the weekdays (mon=1, sun=7)
}
\description{
The function returns a vector of integer, where mondays are 1 and sundays are 7.
}
\note{
Function could have been written with pure strptime commands in case
one does not want the lubridate dependencies.
}

