% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getIncidence.R
\docType{methods}
\name{get.incidence2,obkData-method}
\alias{get.incidence2,obkData-method}
\title{More powerful get.incidence method for obkData}
\usage{
\S4method{get.incidence2}{obkData}(x, data, where = NULL, val.min = NULL,
  val.max = NULL, val.kept = NULL, regexp = NULL, from = NULL,
  to = NULL, interval = 1, add.zero = TRUE, doBy = NULL, ...)
}
\description{
More powerful get.incidence method for obkData
}

